pyproject.toml
[tool.poetry]
name = "simple-flask-api"
version = "0.1.0"
description = ""
authors = ["Cristiano Rocha <sdxadev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
python-dotenv = "^0.20.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
